Update cancellation function in views.py to delete test drive once confirmed by the user

+++ booking/views.py
@@ -36,9 +36,10 @@ def dashboard(request):
 
 def cancellation(request, booking_id):
     booking_cancellation = get_object_or_404(Booking, pk=booking_id)
-    booking_cancellation.delete()
-    messages.add_message(request, messages.INFO, 'Your test drive has been cancelled !!!')
-    return redirect('dashboard')
+    if request.method == 'POST':
+        booking_cancellation.delete()
+        return redirect('dashboard')
+    return render(request, 'booking/booking_cancellation.html')
 
 def edit(request, booking_id):
     booking_id = get_object_or_404(Booking, pk=booking_id)
